Refactor Chat into smaller discreet chunks for better maintainability
