+++ frontend/src/Chat.tsx
@@ -1,6 +1,9 @@
 import { useState, useEffect, useRef } from "react";
 import ExportMessagesButton from "./pages/Chat/components/ExportMessagesButton";
 import { useNavigate } from "react-router-dom";
+import MessageContent from "./pages/Chat/components/MessageContent";
+import MessageFeedback from "./pages/Chat/components/MessageFeedback";
+import InputField from "./pages/Chat/components/InputField";
 
 // Generate a random session ID
 const generateSessionId = (): string => {
@@ -25,12 +28,9 @@ export interface IMessage {
 }
 
 export default function Chat() {
-  const [text, setText] = useState("");
   const [messages, setMessages] = useState<IMessage[]>([]);
   const [isLoading, setIsLoading] = useState(false);
   const [sessionId, setSessionId] = useState<string>("");
-  const [feedbackOpen, setFeedbackOpen] = useState<string | null>(null);
-  const [betterResponse, setBetterResponse] = useState("");
   const [feedbackSubmitted, setFeedbackSubmitted] = useState(false);
   const inputRef = useRef<HTMLInputElement | null>(null);
   const messagesRef = useRef<HTMLDivElement | null>(null);
@@ -85,132 +85,11 @@ export default function Chat() {
     }
   }, []);
 
-  const handleFeedback = async (_messageId: string, betterText: string) => {
-    if (!betterText.trim()) return;
-
-    try {
-      await fetch("/api/feedback", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          session_id: sessionId,
-          comment: betterText,
-        }),
-      });
-
-      // Close feedback form and reset
-      setFeedbackOpen(null);
-      setBetterResponse("");
-      setFeedbackSubmitted(true);
-
-      // Add a refresh message and mark session as feedback submitted
-      setMessages((prev) => [
-        ...prev,
-        {
-          role: "assistant",
-          content:
-            "Thank you for your feedback! Please refresh the page to start a new conversation.",
-          messageId: Date.now().toString(),
-          feedbackSubmitted: true,
-        },
-      ]);
-    } catch (error) {
-      console.error("Error submitting feedback:", error);
-    }
-  };
-
   const handleClearSession = async () => {
     const newSessionId = generateSessionId();
     localStorage.setItem("sessionId", newSessionId);
     setSessionId(newSessionId);
     setMessages([]);
-  };
-
-  const handleSend = async () => {
-    // If feedback was submitted, disable further interaction
-    if (feedbackSubmitted) return;
-
-    if (!text.trim()) return;
-
-    const userMessage = text;
-    const userMessageId = Date.now().toString();
-    const botMessageId = (Date.now() + 1).toString();
-
-    setText("");
-    setIsLoading(true);
-
-    // Add user message
-    setMessages((prev) => [
-      ...prev,
-      { role: "user", content: userMessage, messageId: userMessageId },
-    ]);
-
-    // Add empty bot message that will be updated
-    setMessages((prev) => [
-      ...prev,
-      {
-        role: "assistant",
-        content: "",
-        messageId: botMessageId,
-        showFeedback: false,
-      },
-    ]);
-
-    try {
-      const res = await fetch("/api/query", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ message: userMessage, session_id: sessionId }),
-      });
-
-      const reader = res.body?.getReader();
-      const decoder = new TextDecoder();
-      let fullText = "";
-
-      if (reader) {
-        while (true) {
-          const { done, value } = await reader.read();
-          if (done) break;
-          const chunk = decoder.decode(value);
-          fullText += chunk;
-
-          // Update only the bot's message
-          setMessages((prev) =>
-            prev.map((msg) =>
-              msg.messageId === botMessageId
-                ? { ...msg, content: fullText }
-                : msg
-            )
-          );
-        }
-
-        // Set showFeedback to false for all messages, then true only for the latest bot message
-        setMessages((prev) =>
-          prev.map((msg) =>
-            msg.role === "assistant"
-              ? { ...msg, showFeedback: msg.messageId === botMessageId }
-              : msg
-          )
-        );
-      }
-    } catch (error) {
-      console.error("Error:", error);
-      setMessages((prev) =>
-        prev.map((msg) => {
-          if (msg.messageId === botMessageId) {
-            return {
-              ...msg,
-              content: "Sorry, I encountered an error. Please try again.",
-            };
-          } else if (msg.role === "assistant") {
-            return { ...msg, showFeedback: false };
-          }
-          return msg;
-        })
-      );
-    } finally {
-      setIsLoading(false);
-    }
   };
 
   useEffect(() => {
@@ -258,71 +137,14 @@ export default function Chat() {
                         : "user-message"
                     }`}
                   >
-                    <div className="message-content">
-                      <strong>
-                        {message.role === "assistant" ? "Bot: " : "You: "}
-                      </strong>
-                      {message.role === "assistant" &&
-                      message.content === "" &&
-                      isLoading ? (
-                        <span className="dot-pulse">...</span>
-                      ) : (
-                        <span className="whitespace-pre-wrap">
-                          {message.content}
-                        </span>
-                      )}
-                    </div>
-
+                    <MessageContent message={message} isLoading={isLoading} />
                     {message.role === "assistant" && message.showFeedback && (
-                      <div className="feedback-section">
-                        {message.feedbackSubmitted === true ? (
-                          <div className="feedback-submitted">
-                            <span className="text-green-700">
-                              Thank you for your feedback!
-                            </span>
-                          </div>
-                        ) : feedbackOpen === message.messageId ? (
-                          <div className="feedback-form">
-                            <textarea
-                              className="w-full p-3 border-1 border-[#ddd] rounded-md box-border transition-colors duration-300 focus:outline-0 focus:border-[#4a90e2] focus:shadow-[0_0_0_2px_rgba(74,144,226,0.2)]"
-                              placeholder="Describe the preferred behavior"
-                              value={betterResponse}
-                              onChange={(e) =>
-                                setBetterResponse(e.target.value)
-                              }
-                              rows={4}
-                            />
-                            <div className="flex gap-2">
-                              <button
-                                className="py-1.5 px-4 bg-[#4a90e2] hover:bg-[#3a7bc8] text-white rounded-md cursor-pointer transition-color duration-300"
-                                onClick={() =>
-                                  handleFeedback(
-                                    message.messageId,
-                                    betterResponse
-                                  )
-                                }
-                                disabled={!betterResponse.trim()}
-                              >
-                                Submit
-                              </button>
-                              <button
-                                className="py-1.5 px-4 bg-[#ddd] text-[#333] rounded-md cursor-pointer transition-color duration-300"
-                                onClick={() => setFeedbackOpen(null)}
-                              >
-                                Cancel
-                              </button>
-                            </div>
-                          </div>
-                        ) : (
-                          <button
-                            onClick={() => setFeedbackOpen(message.messageId)}
-                            className="bg-none border-none cursor-pointer p-1 text-[#888]"
-                            title="Provide better response"
-                          >
-                            👎 This response could be better
-                          </button>
-                        )}
-                      </div>
+                      <MessageFeedback
+                        message={message}
+                        setMessages={setMessages}
+                        sessionId={sessionId}
+                        setFeedbackSubmitted={setFeedbackSubmitted}
+                      />
                     )}
                   </div>
                 ))}
@@ -335,34 +157,13 @@ export default function Chat() {
           </div>
         </div>
         <div>
-          <div className="flex gap-2 mt-4 h-11 items-stretch mx-auto max-w-[700px]">
-            <input
-              type="text"
-              value={text}
-              onChange={(e) => setText(e.target.value)}
-              onKeyDown={(e) => {
-                if (e.key === "Enter") {
-                  e.preventDefault(); // prevent form submission or newline
-                  handleSend();
-                }
-              }}
-              className="w-full p-3 border-1 border-[#ddd] rounded-md box-border transition-colors duration-300 focus:outline-0 focus:border-[#4a90e2] focus:shadow-[0_0_0_2px_rgba(74,144,226,0.2)]"
-              placeholder={
-                feedbackSubmitted
-                  ? "Please refresh the page to start a new conversation"
-                  : "Type your message here..."
-              }
-              disabled={isLoading || feedbackSubmitted}
-              ref={inputRef}
-            />
-            <button
-              className="px-6 bg-[#4a90e2] hover:bg-[#3a7bc8] text-white rounded-md cursor-pointer transition-color duration-300"
-              onClick={handleSend}
-              disabled={isLoading || !text.trim() || feedbackSubmitted}
-            >
-              {isLoading ? "..." : "Send"}
-            </button>
-          </div>
+          <InputField
+            setMessages={setMessages}
+            isLoading={isLoading}
+            setIsLoading={setIsLoading}
+            feedbackSubmitted={feedbackSubmitted}
+            sessionId={sessionId}
+          />
           <div className="flex justify-center mt-4">
             <button
               className="cursor-pointer font-bold underline text-[#E3574B] hover:text-[#B8473D]"
